requirements.txt
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
biopython==1.85
scikit-learn==1.6.1
pandas==2.2.3
numpy==1.26.4
matplotlib==3.10.1
seaborn==0.13.2
tqdm==4.67.1
loguru==0.7.3
openpyxl==3.1.5
